requirements.txt
asgiref==3.5.2
Django==4.1.1
django-semantic-admin==0.3.2
djangorestframework==3.13.1
Pillow==9.2.0
pyotp==2.6.0
pytz==2022.2.1
sqlparse==0.4.2
sweetify==2.3.1
